test demand and retain

+++ jetty-websocket/javax-websocket-tests/src/main/java/org/eclipse/jetty/websocket/jsr356/tests/DummyChannel.java
@@ -139,7 +139,7 @@ public class DummyChannel implements FrameHandler.CoreSession
     }
 
     @Override
-    public void demand(int n)
+    public void demand(long n)
     {        
     }
 }
