+++ jetty-websocket/jetty-websocket-common/src/test/java/org/eclipse/jetty/websocket/common/DummyChannel.java
@@ -138,7 +138,7 @@ public class DummyChannel implements FrameHandler.CoreSession
     }
 
     @Override
-    public void demand(int n)
+    public void demand(long n)
     {        
     }
 }
